added zip code mask and update uninstall

+++ uninstall.php
@@ -1,7 +1,10 @@
 <?php
 
 // If uninstall not called from WordPress exit.
-if ( !defined( 'WP_UNINSTALL_PLUGIN' ) )
-    exit ();
+if ( !defined( 'WP_UNINSTALL_PLUGIN' ) || !WP_UNINSTALL_PLUGIN || dirname( WP_UNINSTALL_PLUGIN ) != dirname( plugin_basename( __FILE__ ) ) ) {
+
+    status_header( 404 );
+    exit;
+}
 
 delete_option( 'wcbcf_settings' );
